Initial Typescript conversion
