+++ assets/js/main.js
@@ -1,508 +1,1 @@
-class App {
-  constructor() {
-    this.canCreateMarkers = getNativeFunction("canCreateMarkers");
-    this.createMarkers = getNativeFunction("createMarkers");
-    this.getAudioSources = getNativeFunction("getAudioSources");
-    this.getModels = getNativeFunction("getModels");
-    this.getRegionSequences = getNativeFunction("getRegionSequences");
-    this.getTranscriptionStatus = getNativeFunction("getTranscriptionStatus");
-    this.getWhisperLanguages = getNativeFunction("getWhisperLanguages");
-    this.play = getNativeFunction("play");
-    this.stop = getNativeFunction("stop");
-    this.setPlaybackPosition = getNativeFunction("setPlaybackPosition");
-    this.setWebState = getNativeFunction("setWebState");
-    this.transcribeAudioSource = getNativeFunction("transcribeAudioSource");
-
-    this.state = {
-      modelName: 'small',
-      language: '',
-      translate: false,
-      transcript: null
-    };
-  }
-
-  init() {
-    this.loadState().then(() => {
-      this.initModels();
-      this.initLanguages();
-      this.initTranscript();
-    });
-
-    document.getElementById('process-button').onclick = () => { this.handleProcess(); };
-    document.getElementById('create-markers').onclick = () => { this.handleCreateMarkers('markers'); };
-    document.getElementById('create-regions').onclick = () => { this.handleCreateMarkers('regions'); };
-    document.getElementById('create-notes').onclick = () => { this.handleCreateMarkers('notes'); };
-
-    setInterval(() => {
-      this.update();
-    }, 500);
-  }
-
-  loadState() {
-    try {
-      this.state = JSON.parse(window.__JUCE__.initialisationData.webState);
-    } catch (e) {
-      console.warn('Failed to parse web state:', e);
-    }
-    return Promise.resolve();
-  }
-
-  saveState() {
-    if (this.state) {
-      return this.setWebState(JSON.stringify(this.state));
-    }
-    return Promise.resolve();
-  }
-
-  initModels() {
-    this.getModels().then((models) => {
-      const select = document.getElementById('model-select');
-
-      models.forEach((model) => {
-        const option = document.createElement('option');
-        option.selected = (this.state.modelName === model.name);
-        option.value = model.name;
-        option.innerText = model.label;
-        select.appendChild(option);
-      });
-
-      select.onchange = () => {
-        this.state.modelName = select.options[select.selectedIndex].value;
-        this.saveState();
-      };
-    });
-  }
-
-  initLanguages() {
-    this.getWhisperLanguages().then((languages) => {
-      const select = document.getElementById('language-select');
-
-      languages.forEach((language) => {
-        const option = document.createElement('option');
-        option.selected = (this.state.language === language.code);
-        option.value = language.code;
-        option.innerText = language.name.charAt(0).toUpperCase() + language.name.slice(1);
-        select.appendChild(option);
-      });
-
-      select.onchange = () => {
-        this.state.language = select.options[select.selectedIndex].value;
-        this.saveState();
-      };
-    });
-
-    const translateCheckbox = document.getElementById('translate-checkbox');
-    translateCheckbox.checked = this.state.translate;
-    translateCheckbox.onchange = () => {
-      this.state.translate = translateCheckbox.checked;
-      this.saveState();
-    };
-  }
-
-  initTranscript() {
-    if (this.state.transcript) {
-      const groups = this.state.transcript.groups;
-      if (groups && groups.length > 0) {
-        this.showTranscript();
-        groups.forEach((group) => {
-          this.addSegments(group.segments, group.audioSource);
-        });
-      }
-    }
-  }
-
-  handleProcess() {
-    this.disableProcessButton();
-    this.showSpinner();
-    this.setProcessText('Processing...');
-    this.clearTranscript();
-    this.hideTranscript();
-
-    const languageSelect = document.getElementById('language-select');
-    const languageCode = languageSelect.options[languageSelect.selectedIndex].value;
-    const translate = document.getElementById('translate-checkbox').checked;
-    const asrOptions = {
-      modelName: this.state.modelName,
-      language: languageCode,
-      translate: translate
-    };
-
-    this.getAudioSources().then((audioSources) => {
-      const processNextAudioSource = () => {
-        if (audioSources.length === 0) {
-          this.enableProcessButton();
-          this.hideSpinner();
-          this.setProcessText('Process');
-          this.saveState();
-          return;
-        }
-
-        const audioSource = audioSources.shift();
-
-        this.transcribeAudioSource(audioSource.persistentID, asrOptions).then((result) => {
-          if (result.segments && result.segments.length > 0) {
-            this.showTranscript();
-            this.addSegments(result.segments, audioSource);
-            this.state.transcript = this.state.transcript || { groups: [] };
-            this.state.transcript.groups.push({ segments: result.segments, audioSource: audioSource });
-          } else if (result.error) {
-            this.showAlert('danger', '<b>Error:</b> ' + this.htmlEscape(result.error));
-            audioSources.length = 0;
-          }
-
-          processNextAudioSource();
-        });
-      };
-
-      processNextAudioSource();
-    });
-  }
-
-  handleCreateMarkers(markerType) {
-    const segments = document.querySelectorAll('.segment');
-    let markers = [];
-
-    for (let i = 0; i < segments.length; i++) {
-      const segment = segments[i];
-      const segmentStart = segment.querySelector('.segment-start');
-      const segmentEnd = segment.querySelector('.segment-end');
-
-      if (segmentStart.dataset.playbackStart) {
-        const playbackStart = segmentStart.dataset.playbackStart;
-        const playbackEnd = segmentEnd.dataset.playbackEnd;
-        const text = segment.querySelector('.segment-text').innerText;
-
-        markers.push({
-          start: playbackStart,
-          end: playbackEnd,
-          name: text
-        });
-      }
-    }
-
-    if (markers.length > 0) {
-      this.createMarkers(markers, markerType).then((result) => {
-        if (result && result.error) {
-          this.showAlert('danger', '<b>Error:</b> ' + this.htmlEscape(result.error));
-        }
-      });
-    }
-  }
-
-  update() {
-    this.updateTranscriptionStatus();
-    this.updatePlaybackRegions();
-  }
-
-  updateTranscriptionStatus() {
-    this.getTranscriptionStatus().then((status) => {
-      if (status !== '') {
-        this.setProcessText(status + '...');
-      }
-    });
-  }
-
-  updatePlaybackRegions() {
-    this.getRegionSequences().then((regionSequences) => {
-      this.updatePlaybackForRegionSequences(regionSequences);
-    });
-  }
-
-  updatePlaybackForRegionSequences(regionSequences) {
-    const playbackRegionsByAudioSource =
-      this.collectPlaybackRegionsByAudioSource(regionSequences);
-
-    for (const segment of document.querySelectorAll('.segment')) {
-      const source = segment.querySelector('.segment-source');
-      const sourceID = source.dataset.persistentId;
-      const playbackRegions = playbackRegionsByAudioSource[sourceID] || [];
-
-      this.updatePlaybackForSegment(segment, playbackRegions);
-    }
-  }
-
-  updatePlaybackForSegment(segment, playbackRegions) {
-    const startElement = segment.querySelector('.segment-start');
-    const endElement = segment.querySelector('.segment-end');
-    const textElement = segment.querySelector('.segment-text');
-
-    const segmentStart = parseFloat(startElement.dataset.segmentTime);
-    const segmentEnd = parseFloat(endElement.dataset.segmentTime);
-
-    const playbackRange =
-      this.findPlayableRegion(playbackRegions, segmentStart, segmentEnd);
-
-    if (playbackRange) {
-      startElement.dataset.playbackStart = playbackRange.start;
-      endElement.dataset.playbackEnd = playbackRange.end;
-      textElement.dataset.playbackStart = playbackRange.start;
-      startElement.innerText = this.timestampToString(playbackRange.start);
-      endElement.innerText = this.timestampToString(playbackRange.end);
-    } else {
-      startElement.dataset.playbackStart = '';
-      endElement.dataset.playbackEnd = '';
-      textElement.dataset.playbackStart = '';
-      startElement.innerText = '';
-      endElement.innerText = '';
-    }
-  }
-
-  collectPlaybackRegionsByAudioSource(regionSequences) {
-    const result = {};
-    for (const rs of regionSequences) {
-      for (const pr of rs.playbackRegions) {
-        const sourceID = pr.audioSourcePersistentID;
-        if (!result[sourceID]) {
-          result[sourceID] = [];
-        }
-        result[sourceID].push(pr);
-      }
-    }
-    return result;
-  }
-
-  findPlayableRegion(playbackRegions, segmentStart, segmentEnd) {
-    for (const pr of playbackRegions) {
-      const playbackStart = pr.playbackStart;
-      const playbackEnd = pr.playbackEnd;
-      const modificationStart = pr.modificationStart;
-
-      const start = playbackStart + segmentStart - modificationStart;
-      const end = playbackStart + segmentEnd - modificationStart;
-
-      if (start >= playbackStart && start <= playbackEnd) {
-        return { start, end };
-      }
-    }
-
-    return null;
-  }
-
-  enableProcessButton() {
-    document.getElementById('process-button').disabled = false;
-  }
-
-  disableProcessButton() {
-    document.getElementById('process-button').disabled = true;
-  }
-
-  setProcessText(text) {
-    document.getElementById('process-text').innerText = text;
-  }
-
-  showSpinner() {
-    document.getElementById('spinner').style.display = 'inline-block';
-  }
-
-  hideSpinner() {
-    document.getElementById('spinner').style.display = 'none';
-  }
-
-  showAlert(type, message) {
-    const alerts = document.getElementById('alerts');
-    const wrapper = document.createElement('div');
-    wrapper.innerHTML = [
-      `<div class="alert alert-${type} alert-dismissible mb-2" role="alert">`,
-      `   <div>${message}</div>`,
-      '   <button type="button" class="btn-close" data-bs-dismiss="alert" aria-label="Close"></button>',
-      '</div>'
-    ].join('');
-    alerts.append(wrapper);
-  }
-
-  showCreateMenu() {
-    document.getElementById('create-menu').style.display = 'block';
-  }
-
-  showTranscript() {
-    document.getElementById('transcript').style.display = 'block';
-
-    this.canCreateMarkers().then((canCreateMarkers) => {
-      if (canCreateMarkers) {
-        this.showCreateMenu();
-      }
-    });
-  }
-
-  hideTranscript() {
-    document.getElementById('transcript').style.display = 'none';
-  }
-
-  clearTranscript() {
-    const table = document.querySelector('#transcript table');
-    const tbody = table.querySelector('tbody');
-    tbody.innerHTML = '';
-
-    this.state.transcript = null;
-    this.saveState();
-  }
-
-  addSegments(segments, audioSource) {
-    this.addRows(segments.map((segment) => [
-      this.formatStartTime(segment.start, 'segment-start small'),
-      this.formatEndTime(segment.end, 'segment-end small text-muted'),
-      this.formatText(segment.text, 'segment-text'),
-      this.formatScore(segment, 'segment-score'),
-      this.formatSource(audioSource.name, 'segment-source small', audioSource.persistentID),
-    ]));
-  }
-
-  addRows(rows) {
-    const table = document.querySelector('#transcript table');
-    const tbody = table.querySelector('tbody');
-    const fragment = document.createDocumentFragment();
-
-    rows.forEach(row => {
-      const tr = document.createElement('tr');
-      tr.className = 'segment align-middle';
-
-      const cells = row.map((cell, index) => {
-        const td = document.createElement('td');
-        td.innerHTML = cell;
-
-        if (index === 4) {
-          // Truncate audio source name with tooltip
-          td.className = 'text-muted text-truncate';
-          td.style = 'max-width: 200px;';
-          td.title = td.textContent;
-        }
-
-        return td;
-      });
-
-      tr.append(...cells);
-      fragment.appendChild(tr);
-    });
-
-    tbody.appendChild(fragment);
-  }
-
-  formatStartTime(segmentTime, className) {
-    const linkClasses = 'link-offset-2 link-underline link-underline-opacity-0 link-underline-opacity-50-hover';
-    return `<a href="javascript:" onclick="app.playSegment(this)" class="${className} ${linkClasses}" data-segment-time="${segmentTime}">${this.timestampToString(segmentTime)}</a>`;
-  }
-
-  formatEndTime(segmentTime, className) {
-    return `<span class="${className}" data-segment-time="${segmentTime}">${this.timestampToString(segmentTime)}</span>`;
-  }
-
-  formatText(text, className) {
-    const linkClasses = 'link-light link-offset-2 link-underline link-underline-opacity-0 link-underline-opacity-50-hover';
-    return `<a href="javascript:" onclick="app.playSegment(this)" class="${className} ${linkClasses}">${this.htmlEscape(text)}</a>`;
-  }
-
-  formatScore(segment, className) {
-    const score = segment.score;
-    const color = this.scoreColor(score);
-    const percentage = score * 100;
-    return [
-      '<div class="progress" style="height: 2px">',
-      `  <div class="progress-bar ${className}" style="width: ${percentage}%; background-color: ${color}"></div>`,
-      '</div>'
-    ].join('');
-  }
-
-  formatSource(text, className, audioSourcePersistentID) {
-    return `<span class="${className}" data-persistent-id="${this.htmlEscape(audioSourcePersistentID)}">${this.htmlEscape(text)}</span>`;
-  }
-
-  htmlEscape(str) {
-    return str.replace(/&/g, '&amp;').replace(/</g, '&lt;').replace(/>/g, '&gt;').replace(/"/g, '&quot;');
-  }
-
-  playSegment(segment) {
-    if (segment.dataset.playbackStart) {
-      const playbackStart = parseFloat(segment.dataset.playbackStart);
-      this.playAt(playbackStart);
-    }
-  }
-
-  playAt(seconds) {
-    this.stop();
-    this.setPlaybackPosition(seconds).then(() => {
-      this.play();
-    });
-  }
-
-  scoreColor(value) {
-    if (value > 0.9) {
-      return "#a3ff00";
-    } else if (value > 0.8) {
-      return "#2cba00";
-    } else if (value > 0.7) {
-      return "#ffa700";
-    } else if (value > 0.0) {
-      return "#ff2c2f";
-    } else {
-      return "transparent";
-    }
-  }
-
-  // Return a string of the form M:SS.mmm
-  timestampToString(seconds) {
-    const minutes = Math.floor(seconds / 60);
-    const remainingSeconds = seconds - minutes * 60;
-    const milliseconds = Math.round((remainingSeconds - Math.floor(remainingSeconds)) * 1000);
-    return `${minutes}:${String(Math.floor(remainingSeconds)).padStart(2, '0')}.${String(milliseconds).padStart(3, '0')}`;
-  }
-}
-
-if (typeof window.__JUCE__ === "undefined") {
-  function getNativeFunction() {
-    return function() {
-      return new Promise((resolve, reject) => {});
-    };
-  }
-} else {
-  // The following code is copied from modules/juce_gui_extra/native/javascript/index.js
-  // That file is intended to be used with a JS build system, but since we do not
-  // currently use one, this is the easiest way to get native functions to work.
-
-  class PromiseHandler {
-    constructor() {
-      this.lastPromiseId = 0;
-      this.promises = new Map();
-
-      window.__JUCE__.backend.addEventListener(
-        "__juce__complete",
-        ({ promiseId, result }) => {
-          if (this.promises.has(promiseId)) {
-            this.promises.get(promiseId).resolve(result);
-            this.promises.delete(promiseId);
-          }
-        }
-      );
-    }
-
-    createPromise() {
-      const promiseId = this.lastPromiseId++;
-      const result = new Promise((resolve, reject) => {
-        this.promises.set(promiseId, { resolve: resolve, reject: reject });
-      });
-      return [promiseId, result];
-    }
-  }
-
-  const promiseHandler = new PromiseHandler();
-
-  function getNativeFunction(name) {
-    if (!window.__JUCE__.initialisationData.__juce__functions.includes(name))
-      console.warn(
-        `Creating native function binding for '${name}', which is unknown to the backend`
-      );
-
-    const f = function () {
-      const [promiseId, result] = promiseHandler.createPromise();
-
-      window.__JUCE__.backend.emitEvent("__juce__invoke", {
-        name: name,
-        params: Array.prototype.slice.call(arguments),
-        resultId: promiseId,
-      });
-
-      return result;
-    };
-
-    return f;
-  }
-}
+(()=>{var S=Object.create;var y=Object.defineProperty;var k=Object.getOwnPropertyDescriptor;var C=Object.getOwnPropertyNames;var T=Object.getPrototypeOf,x=Object.prototype.hasOwnProperty;var L=(a,e)=>()=>(e||a((e={exports:{}}).exports,e),e.exports);var I=(a,e,t,s)=>{if(e&&typeof e=="object"||typeof e=="function")for(let n of C(e))!x.call(a,n)&&n!==t&&y(a,n,{get:()=>e[n],enumerable:!(s=k(e,n))||s.enumerable});return a};var J=(a,e,t)=>(t=a!=null?S(T(a)):{},I(e||!a||!a.__esModule?y(t,"default",{value:a,enumerable:!0}):t,a));var b=L((exports,module)=>{typeof window.__JUCE__<"u"&&typeof window.__JUCE__.getAndroidUserScripts<"u"&&typeof window.inAndroidUserScriptEval>"u"&&(window.inAndroidUserScriptEval=!0,eval(window.__JUCE__.getAndroidUserScripts()),delete window.inAndroidUserScriptEval);{typeof window.__JUCE__>"u"&&(console.warn("The 'window.__JUCE__' object is undefined. Native integration features will not work. Defining a placeholder 'window.__JUCE__' object."),window.__JUCE__={postMessage:function(){}}),typeof window.__JUCE__.initialisationData>"u"&&(window.__JUCE__.initialisationData={__juce__platform:[],__juce__functions:[],__juce__registeredGlobalEventIds:[],__juce__sliders:[],__juce__toggles:[],__juce__comboBoxes:[]});class a{constructor(){this.listeners=new Map,this.listenerId=0}addListener(n){let i=this.listenerId++;return this.listeners.set(i,n),i}removeListener(n){this.listeners.has(n)&&this.listeners.delete(n)}callListeners(n){for(let[,i]of this.listeners)i(n)}}class e{constructor(){this.eventListeners=new Map}addEventListener(n,i){this.eventListeners.has(n)||this.eventListeners.set(n,new a);let o=this.eventListeners.get(n).addListener(i);return[n,o]}removeEventListener([n,i]){this.eventListeners.has(n)&&this.eventListeners.get(n).removeListener(i)}emitEvent(n,i){this.eventListeners.has(n)&&this.eventListeners.get(n).callListeners(i)}}class t{constructor(){this.listeners=new e}addEventListener(n,i){return this.listeners.addEventListener(n,i)}removeEventListener([n,i]){this.listeners.removeEventListener(n,i)}emitEvent(n,i){window.__JUCE__.postMessage(JSON.stringify({eventId:n,payload:i}))}emitByBackend(n,i){this.listeners.emitEvent(n,JSON.parse(i))}}typeof window.__JUCE__.backend>"u"&&(window.__JUCE__.backend=new t)}});var D=J(b()),_=class{constructor(){this.lastPromiseId=0,this.promises=new Map,window.__JUCE__.backend.addEventListener("__juce__complete",({promiseId:e,result:t})=>{this.promises.has(e)&&(this.promises.get(e).resolve(t),this.promises.delete(e))})}createPromise(){let e=this.lastPromiseId++,t=new Promise((s,n)=>{this.promises.set(e,{resolve:s,reject:n})});return[e,t]}},M=new _;function l(a){return window.__JUCE__.initialisationData.__juce__functions.includes(a)||console.warn(`Creating native function binding for '${a}', which is unknown to the backend`),function(){let[t,s]=M.createPromise();return window.__JUCE__.backend.emitEvent("__juce__invoke",{name:a,params:Array.prototype.slice.call(arguments),resultId:t}),s}}var u=class{constructor(){this.listeners=new Map,this.listenerId=0}addListener(e){let t=this.listenerId++;return this.listeners.set(t,e),t}removeListener(e){this.listeners.has(e)&&this.listeners.delete(e)}callListeners(e){for(let[,t]of this.listeners)t(e)}},h="valueChanged",E="propertiesChanged",U="sliderDragStarted",P="sliderDragEnded",g=class{constructor(e){window.__JUCE__.initialisationData.__juce__sliders.includes(e)||console.warn("Creating SliderState for '"+e+"', which is unknown to the backend"),this.name=e,this.identifier="__juce__slider"+this.name,this.scaledValue=0,this.properties={start:0,end:1,skew:1,name:"",label:"",numSteps:100,interval:0,parameterIndex:-1},this.valueChangedEvent=new u,this.propertiesChangedEvent=new u,window.__JUCE__.backend.addEventListener(this.identifier,t=>this.handleEvent(t)),window.__JUCE__.backend.emitEvent(this.identifier,{eventType:"requestInitialUpdate"})}setNormalisedValue(e){this.scaledValue=this.snapToLegalValue(this.normalisedToScaledValue(e)),window.__JUCE__.backend.emitEvent(this.identifier,{eventType:h,value:this.scaledValue})}sliderDragStarted(){window.__JUCE__.backend.emitEvent(this.identifier,{eventType:U})}sliderDragEnded(){window.__JUCE__.backend.emitEvent(this.identifier,{eventType:P})}handleEvent(e){if(e.eventType==h&&(this.scaledValue=e.value,this.valueChangedEvent.callListeners()),e.eventType==E){let{eventType:t,...s}=e;this.properties=s,this.propertiesChangedEvent.callListeners()}}getScaledValue(){return this.scaledValue}getNormalisedValue(){return Math.pow((this.scaledValue-this.properties.start)/(this.properties.end-this.properties.start),this.properties.skew)}normalisedToScaledValue(e){return Math.pow(e,1/this.properties.skew)*(this.properties.end-this.properties.start)+this.properties.start}snapToLegalValue(e){let t=this.properties.interval;if(t==0)return e;let s=this.properties.start;return((i,o=0,c=1)=>Math.max(o,Math.min(c,i)))(s+t*Math.floor((e-s)/t+.5),this.properties.start,this.properties.end)}},B=new Map;for(let a of window.__JUCE__.initialisationData.__juce__sliders)B.set(a,new g(a));var f=class{constructor(e){window.__JUCE__.initialisationData.__juce__toggles.includes(e)||console.warn("Creating ToggleState for '"+e+"', which is unknown to the backend"),this.name=e,this.identifier="__juce__toggle"+this.name,this.value=!1,this.properties={name:"",parameterIndex:-1},this.valueChangedEvent=new u,this.propertiesChangedEvent=new u,window.__JUCE__.backend.addEventListener(this.identifier,t=>this.handleEvent(t)),window.__JUCE__.backend.emitEvent(this.identifier,{eventType:"requestInitialUpdate"})}getValue(){return this.value}setValue(e){this.value=e,window.__JUCE__.backend.emitEvent(this.identifier,{eventType:h,value:this.value})}handleEvent(e){if(e.eventType==h&&(this.value=e.value,this.valueChangedEvent.callListeners()),e.eventType==E){let{eventType:t,...s}=e;this.properties=s,this.propertiesChangedEvent.callListeners()}}},F=new Map;for(let a of window.__JUCE__.initialisationData.__juce__toggles)F.set(a,new f(a));var v=class{constructor(e){window.__JUCE__.initialisationData.__juce__comboBoxes.includes(e)||console.warn("Creating ComboBoxState for '"+e+"', which is unknown to the backend"),this.name=e,this.identifier="__juce__comboBox"+this.name,this.value=0,this.properties={name:"",parameterIndex:-1,choices:[]},this.valueChangedEvent=new u,this.propertiesChangedEvent=new u,window.__JUCE__.backend.addEventListener(this.identifier,t=>this.handleEvent(t)),window.__JUCE__.backend.emitEvent(this.identifier,{eventType:"requestInitialUpdate"})}getChoiceIndex(){return Math.round(this.value*(this.properties.choices.length-1))}setChoiceIndex(e){let t=this.properties.choices.length;this.value=t>1?e/(t-1):0,window.__JUCE__.backend.emitEvent(this.identifier,{eventType:h,value:this.value})}handleEvent(e){if(e.eventType==h&&(this.value=e.value,this.valueChangedEvent.callListeners()),e.eventType==E){let{eventType:t,...s}=e;this.properties=s,this.propertiesChangedEvent.callListeners()}}},A=new Map;for(let a of window.__JUCE__.initialisationData.__juce__comboBoxes)A.set(a,new v(a));var w=class{canCreateMarkers=l("canCreateMarkers");createMarkers=l("createMarkers");getAudioSources=l("getAudioSources");getModels=l("getModels");getRegionSequences=l("getRegionSequences");getTranscriptionStatus=l("getTranscriptionStatus");getWhisperLanguages=l("getWhisperLanguages");play=l("play");stop=l("stop");setPlaybackPosition=l("setPlaybackPosition");setWebState=l("setWebState");transcribeAudioSource=l("transcribeAudioSource")},m=class{native;state;constructor(e){this.native=e||new w,this.state={modelName:"small",language:"",translate:!1,transcript:null}}init(){this.loadState().then(()=>{this.initModels(),this.initLanguages(),this.initTranscript()}),document.getElementById("process-button").onclick=()=>{this.handleProcess()},document.getElementById("create-markers").onclick=()=>{this.handleCreateMarkers("markers")},document.getElementById("create-regions").onclick=()=>{this.handleCreateMarkers("regions")},document.getElementById("create-notes").onclick=()=>{this.handleCreateMarkers("notes")},setInterval(()=>{this.update()},500)}loadState(){if(!window.__JUCE__.initialisationData.webState)return Promise.resolve();try{this.state=JSON.parse(window.__JUCE__.initialisationData.webState)}catch(e){console.warn("Failed to parse web state:",e)}return Promise.resolve()}saveState(){return this.state?this.native.setWebState(JSON.stringify(this.state)):Promise.resolve()}initModels(){this.native.getModels().then(e=>{let t=document.getElementById("model-select");e.forEach(s=>{let n=document.createElement("option");n.selected=this.state.modelName===s.name,n.value=s.name,n.innerText=s.label,t.appendChild(n)}),t.onchange=()=>{this.state.modelName=t.options[t.selectedIndex].value,this.saveState()}})}initLanguages(){this.native.getWhisperLanguages().then(t=>{let s=document.getElementById("language-select");t.forEach(n=>{let i=document.createElement("option");i.selected=this.state.language===n.code,i.value=n.code,i.innerText=n.name.charAt(0).toUpperCase()+n.name.slice(1),s.appendChild(i)}),s.onchange=()=>{this.state.language=s.options[s.selectedIndex].value,this.saveState()}});let e=document.getElementById("translate-checkbox");e.checked=this.state.translate,e.onchange=()=>{this.state.translate=e.checked,this.saveState()}}initTranscript(){if(this.state.transcript){let e=this.state.transcript.groups;e&&e.length>0&&(this.showTranscript(),e.forEach(t=>{this.addSegments(t.segments,t.audioSource)}))}}handleProcess(){this.disableProcessButton(),this.showSpinner(),this.setProcessText("Processing..."),this.clearTranscript(),this.hideTranscript();let e=document.getElementById("language-select"),t=e.options[e.selectedIndex].value,s=document.getElementById("translate-checkbox").checked,n={modelName:this.state.modelName,language:t,translate:s};this.native.getAudioSources().then(i=>{let o=()=>{if(i.length===0){this.enableProcessButton(),this.hideSpinner(),this.setProcessText("Process"),this.saveState();return}let c=i.shift();this.native.transcribeAudioSource(c.persistentID,n).then(r=>{r.segments&&r.segments.length>0?(this.showTranscript(),this.addSegments(r.segments,c),this.state.transcript=this.state.transcript||{groups:[]},this.state.transcript.groups.push({segments:r.segments,audioSource:c})):r.error&&(this.showAlert("danger","<b>Error:</b> "+this.htmlEscape(r.error)),i.length=0),o()})};o()})}handleCreateMarkers(e){let t=document.querySelectorAll(".segment"),s=[];for(let n=0;n<t.length;n++){let i=t[n],o=i.querySelector(".segment-start"),c=i.querySelector(".segment-end");if(o.dataset.playbackStart){let r=o.dataset.playbackStart,p=c.dataset.playbackEnd,d=i.querySelector(".segment-text").innerText;s.push({start:r,end:p,name:d})}}s.length>0&&this.native.createMarkers(s,e).then(n=>{n&&n.error&&this.showAlert("danger","<b>Error:</b> "+this.htmlEscape(n.error))})}update(){this.updateTranscriptionStatus(),this.updatePlaybackRegions()}updateTranscriptionStatus(){this.native.getTranscriptionStatus().then(e=>{e!==""&&this.setProcessText(e+"...")})}updatePlaybackRegions(){this.native.getRegionSequences().then(e=>{this.updatePlaybackForRegionSequences(e)})}updatePlaybackForRegionSequences(e){let t=this.collectPlaybackRegionsByAudioSource(e);for(let s of document.querySelectorAll(".segment")){let i=s.querySelector(".segment-source").dataset.persistentId,o=t[i]||[];this.updatePlaybackForSegment(s,o)}}updatePlaybackForSegment(e,t){let s=e.querySelector(".segment-start"),n=e.querySelector(".segment-end"),i=e.querySelector(".segment-text"),o=parseFloat(s.dataset.segmentTime),c=parseFloat(n.dataset.segmentTime),r=this.findPlayableRegion(t,o,c);r?(s.dataset.playbackStart=r.start,n.dataset.playbackEnd=r.end,i.dataset.playbackStart=r.start,s.innerText=this.timestampToString(r.start),n.innerText=this.timestampToString(r.end)):(s.dataset.playbackStart="",n.dataset.playbackEnd="",i.dataset.playbackStart="",s.innerText="",n.innerText="")}collectPlaybackRegionsByAudioSource(e){let t={};for(let s of e)for(let n of s.playbackRegions){let i=n.audioSourcePersistentID;t[i]||(t[i]=[]),t[i].push(n)}return t}findPlayableRegion(e,t,s){for(let n of e){let i=n.playbackStart,o=n.playbackEnd,c=n.modificationStart,r=i+t-c,p=i+s-c;if(r>=i&&r<=o)return{start:r,end:p}}return null}enableProcessButton(){document.getElementById("process-button").disabled=!1}disableProcessButton(){document.getElementById("process-button").disabled=!0}setProcessText(e){document.getElementById("process-text").innerText=e}showSpinner(){document.getElementById("spinner").style.display="inline-block"}hideSpinner(){document.getElementById("spinner").style.display="none"}showAlert(e,t){let s=document.getElementById("alerts"),n=document.createElement("div");n.innerHTML=[`<div class="alert alert-${e} alert-dismissible mb-2" role="alert">`,`   <div>${t}</div>`,'   <button type="button" class="btn-close" data-bs-dismiss="alert" aria-label="Close"></button>',"</div>"].join(""),s.append(n)}showCreateMenu(){document.getElementById("create-menu").style.display="block"}showTranscript(){document.getElementById("transcript").style.display="block",this.native.canCreateMarkers().then(e=>{e&&this.showCreateMenu()})}hideTranscript(){document.getElementById("transcript").style.display="none"}clearTranscript(){let t=document.querySelector("#transcript table").querySelector("tbody");t.innerHTML="",this.state.transcript=null,this.saveState()}addSegments(e,t){this.addRows(e.map(s=>[this.formatStartTime(s.start,"segment-start small"),this.formatEndTime(s.end,"segment-end small text-muted"),this.formatText(s.text,"segment-text"),this.formatScore(s,"segment-score"),this.formatSource(t.name,"segment-source small",t.persistentID)]))}addRows(e){let s=document.querySelector("#transcript table").querySelector("tbody"),n=document.createDocumentFragment();e.forEach(i=>{let o=document.createElement("tr");o.className="segment align-middle";let c=i.map((r,p)=>{let d=document.createElement("td");return d.innerHTML=r,p===4&&(d.className="text-muted text-truncate",d.style="max-width: 200px;",d.title=d.textContent),d});o.append(...c),n.appendChild(o)}),s.appendChild(n)}formatStartTime(e,t){return`<a href="javascript:" onclick="app.playSegment(this)" class="${t} link-offset-2 link-underline link-underline-opacity-0 link-underline-opacity-50-hover" data-segment-time="${e}">${this.timestampToString(e)}</a>`}formatEndTime(e,t){return`<span class="${t}" data-segment-time="${e}">${this.timestampToString(e)}</span>`}formatText(e,t){return`<a href="javascript:" onclick="app.playSegment(this)" class="${t} link-light link-offset-2 link-underline link-underline-opacity-0 link-underline-opacity-50-hover">${this.htmlEscape(e)}</a>`}formatScore(e,t){let s=e.score,n=this.scoreColor(s),i=s*100;return['<div class="progress" style="height: 2px">',`  <div class="progress-bar ${t}" style="width: ${i}%; background-color: ${n}"></div>`,"</div>"].join("")}formatSource(e,t,s){return`<span class="${t}" data-persistent-id="${this.htmlEscape(s)}">${this.htmlEscape(e)}</span>`}htmlEscape(e){return e.replace(/&/g,"&amp;").replace(/</g,"&lt;").replace(/>/g,"&gt;").replace(/"/g,"&quot;")}playSegment(e){if(e.dataset.playbackStart){let t=parseFloat(e.dataset.playbackStart);this.playAt(t)}}playAt(e){this.native.stop(),this.native.setPlaybackPosition(e).then(()=>{this.native.play()})}scoreColor(e){return e>.9?"#a3ff00":e>.8?"#2cba00":e>.7?"#ffa700":e>0?"#ff2c2f":"transparent"}timestampToString(e){let t=Math.floor(e/60),s=e-t*60,n=Math.round((s-Math.floor(s))*1e3);return`${t}:${String(Math.floor(s)).padStart(2,"0")}.${String(n).padStart(3,"0")}`}};window.App=m;})();
