+++ source/ts/package.json
@@ -1,0 +1,16 @@
+{
+  "name": "reaspeech-lite",
+  "version": "1.0.0",
+  "description": "",
+  "dependencies": {
+    "juce-framework-frontend": "file:../../JUCE/modules/juce_gui_extra/native/javascript"
+  },
+  "scripts": {
+    "build": "node build.js",
+    "watch": "node build.js --watch"
+  },
+  "devDependencies": {
+    "esbuild": "^0.25.1",
+    "typescript": "^5.0.0"
+  }
+}
