+++ source/ts/tsconfig.json
@@ -1,0 +1,9 @@
+{
+  "compilerOptions": {
+    "target": "es2022",
+    "module": "es2022",
+    "outDir": "../../assets/js",
+    "noImplicitAny": false
+  },
+  "include": ["main.ts"]
+}
